+++ evals/__init__.py
@@ -1,0 +1,1 @@
+"""Esperimenti ripetibili, separati dai test di regressione."""